requirements.txt
arize-phoenix
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation
openinference-instrumentation-anthropic
anthropic
